Support WOF's inception, cessation and deprecated fields.
This parses these fields as EDTF, which is brought in as an
additional library dependency, and includes the start and end
dates in the database rows for each entry. The visibility is
tracked and updated, generating expiries to be re-rendered, which
allows upstream code to only check the visibility parameter and
not need to worry about the dates changing.