Don't pass tile coordinate to post-process functions.
This is laying the groundwork for the zoom level and geographic coordinates to vary independently from one another. The unpadded coordinates are already stored in the post processing context, so we just need to pass in the nominal zoom (i.e: a measure of display scale) rather than the actual tile coordinate to ensure that post-processing steps are not dependent on the coordinate.