Trap MemoryError, print a log message, and halt
MemoryError sometimes gets thrown when a very, very large message is put in a multiprocessing queue. This exception propagates up to the root of the thread and kills it, locking up the worker.

According to the Python docs, MemoryError isn't entirely safe to recover from, so instead this halts the whole process. It will be restarted by runit / systemd, and it might lose some work, but any lost jobs will be re-queued by SQS after a timeout.