[WIP] Fixture-based data source
This adds a fixture-based data source, which returns data in a similar format to the data coming from PostgreSQL, so that it can be processed in the same way by the rest of the system. The idea is to use it for tests before widening the implementation to include RAWR tiles.