Add support for local storage of RAWR tiles.
Both for reading and writing. This patch means that it's possible to configure a store to round-trip RAWR tiles locally. For example, with the following RAWR config:

```yaml
rawr:
  group-zoom: 10
  queue:
    type: mem
    name: myrawrqueue
    input-file: rawr_input.txt # this just contains 10/163/395
  postgresql:
    host: null
    port: 5432
    dbname: california
    user: matt
    password: null
  store:
    type: directory
    path: rawr_tiles
  source:
    type: store
    store:
      type: directory
      name: rawr_tiles
    postgresql:
      host: null
      port: null
      dbname: california
      user: matt
    table-sources:
      planet_osm_line: &osm { name: osm, value: openstreetmap.org }
      planet_osm_point: *osm
      planet_osm_polygon: *osm
      planet_osm_ways: *osm
      planet_osm_rels: *osm
use-rawr-tiles: true
```

I'm able to run:

```
tilequeue rawr-process --config config.yaml
tilequeue tile --config config.yaml 16/10475/25320
```

And generate an empty tile. Intentionally; that tile has nothing in it - easier to check that for correctness than a few pages of GeoJSON!