Handle missing nodes, ways and relations in the index.
This can happen because we don't have perfect information about bi-directional linkages when building the index and, rather than index every element, the index is selective about what it stores. When the selection doesn't anticipate relation membership, it can result in a reference to a non-indexed element. At the moment, I don't think this is a problem, as those elements wouldn't have been "interesting" anyway.