Simplify coord -> queue msg handle mappings
* Instead of using memory locations as the ids, use the coordinate
triple and queue message handle. This makes it more transparent when
inconsistencies occur, and using the queue message handle as its id removes the
need for a separate mapping.
* Update _reject_coord to acknowledge the message now. Previously it
only updated the tracker but wouldn't actually acknowledge to the queue.
This code path shouldn't be taken anyway, only for coordinates higher
than max zoom which shouldn't get enqueued any more.