Filter to different queue based on TOI membership.
This adds the ability to filter to a different queue depending on whether the coordinate is part of the TOI or not. This means that we can configure different setups, such as:

1. Queue filter by zoom, intersect on TOI: Current configuration, renders the TOI only.
2. Queue filter by TOI, intersect all tiles: "Global build" configuration, renders all tiles, but can send TOI tiles to a higher priority queue.

The configuration should allow flexible priorities, for example whether all TOI comes before all non-TOI, or whether they're interleaved by zoom.