Relax precision of floating point comparison in test.
Comparing exactly equal was giving the error message below, which differs only in the 15th decimal place. This is a difference of a fraction of a micrometer, so I think we can call it "good enough". Sadly, there doesn't seem to be an `assertAlmostEqual` which recurses as `assertEqual` does, so the comparison needed to be expanded.

```
======================================================================
FAIL: test_convert_coord (tests.test_tile.TestCoordToBounds)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "tilequeue/tests/test_tile.py", line 54, in test_convert_coord
    self.assertEqual(exp_bounds, bounds)
AssertionError: Tuples differ: (-74.00390625, 40.697299008636... != (-74.00390625, 40.697299008636...

First differing element 3:
40.713955826286046
40.71395582628604

- (-74.00390625, 40.69729900863674, -73.98193359375, 40.713955826286046)
?                                                                     -

+ (-74.00390625, 40.69729900863674, -73.98193359375, 40.71395582628604)

----------------------------------------------------------------------
```