Fix definition of overlaps for boundaries
For boundary geometries, we want to include the row in the result of a query when some part of its boundary is inside the bounding box. Originally, we used `st_intersects` for this, but it had the problem that it would include rows which had very large (and complex) boundaries for tiles which were completely in the interior of the polygon, which led to unnecessary data being returned and processed.

In https://github.com/tilezen/vector-datasource/pull/1116, a new `bbox_overlaps` test was added, using `st_overlaps`. This no longer included polygons for which the tile bounding box was entirely with the interior. However, it also didn't include polygons which were themselves entirely within the interior of the bounding box, which meant some small countries (e.g: Liechtenstein) went missing at some zoom levels!

This patch changes the definition of the `bbox_overlaps` test to also include polygons contained by the tile bounding box.