Add log function to post-process Context
This will allow output of JSON log messages from post-processing functions. The function can be called with any data which is JSON-ifiable, i.e: consists of plain `dict`, `list`, etc... On its way up back out of the call stack it will automatically be annotated with various bits of data about the coordinate being processed and the post-processing step.

The idea is that this gets output to the Batch logs and we can process it from there to analyse further.