Add utility function to orient MultiPolygons.
In the boundary code, we ensure that admin polygons are correctly oriented, with the interior on the left of their rings, before extracting those rings as boundary lines. However, the orient function from Shapely only handles Polygons, not MultiPolygons, and bails with an exception if it encounters a MultiPolygon. Sadly, many admin areas are indeed MultiPolygons. This adds a simple utility function to orient MultiPolygons by orienting all constituent Polygons.