Alternate viewpoints changes for RAWR tiles.
In tilezen/vector-datasource#1895, I added a little hack to work around the way we change the geometry type of `boundaries` layer features. They used to come entirely from the `planet_osm_polygon` table, but now we also include disputes and claims from `planet_osm_line`. However, we have to filter out a bunch of features from `planet_osm_line` that we don't want, but have the same tags as the polygons we do want.

Long story short: The query in `vector-datasource` sets this flag when transforming a polygon into a boundary line, so we have to do the same thing when reading from the RAWR tiles.